hmm, need to be careful with varargs, maybe they aren't such a good thing

+++ src/simpleann/simpleann.c
@@ -27,7 +27,7 @@ void simple_net_new( simple_net *net, transfer tr, size_t num_layers, ... ) {
 
     for ( i = 0; i < num_layers; i++ ) {
 
-        size = va_arg( layer_size, size_t );
+        size = va_arg( layer_size, int );
 
         net->layer_sizes = realloc( net->layer_sizes, sizeof( size_t ) * (i+1) );
         net->layer_sizes[i] = size;
